+++ utils/web.py
@@ -9,7 +9,7 @@ import traceback
 from collections import OrderedDict, namedtuple
 from urllib.parse import urlparse, parse_qs
 from time import sleep
-from typing import Any, NamedTuple
+from typing import Any, NamedTuple, Optional
 
 from textio import print_error, print_warning
 
@@ -137,6 +137,79 @@ def open_url(url_to_open: str) -> None:
 
 def open_get_started_url() -> None:
     open_url('https://github.com/prof79/fansly-downloader-ng/wiki/Getting-Started')
+
+
+def guess_check_key(
+            main_js_pattern: str,
+            check_key_pattern: str,
+            user_agent: str,
+        ) -> Optional[str]:
+    """Tries to guess the check key from the Fansly homepage.
+    
+    :param main_js_pattern: A regular expression to locate the main.*.js file.
+    :type main_js_pattern: str
+
+    :param check_key_pattern: A regular expression to parse the check key.
+    :type check_key_pattern: str
+
+    :param user_agent: Browser user agent to use for requests.
+    :type user_agent: str
+    
+    :return: The check key string if found or None otherwise.
+    :rtype: Optional[str]
+    """
+    fansly_url = 'https://fansly.com'
+
+    headers = {
+        'User-Agent': user_agent,
+    }
+
+    try:
+        html_response = requests.get(
+            fansly_url,
+            headers=headers,
+        )
+
+        if html_response.status_code == 200:
+
+            if html_response.text:
+                main_js_match = re.search(
+                    pattern=main_js_pattern,
+                    string=html_response.text,
+                    flags=re.IGNORECASE | re.MULTILINE,
+                )
+
+                if main_js_match:
+
+                    main_js = main_js_match.group(1)
+
+                    main_js_url = f'{fansly_url}/{main_js}'
+
+                    js_response = requests.get(
+                        main_js_url,
+                        headers=headers,
+                    )
+
+                    if js_response.status_code == 200:
+
+                        if js_response.text:
+
+                            check_key_match = re.search(
+                                pattern=check_key_pattern,
+                                string=js_response.text,
+                                flags=re.IGNORECASE | re.MULTILINE,
+                            )
+
+                            if check_key_match:
+
+                                check_key = check_key_match.group(1)
+
+                                return check_key
+
+    except:
+        pass
+
+    return None
 
 
 def guess_user_agent(user_agents: dict, based_on_browser: str, default_ua: str) -> str: